Allow specifying locking functions for multi-threading on open / create
Having separate set_locking_funcs functions causes the locking functions
to be available potentially too late. And it just is not very pretty.
This commit changes the API (in a way which is backward compat with 0.3.1), so
that locking functions can be specified at creation time, just like all the
other callbacks.

This commit also bumps the reported version to 0.3.2, so that apps using
libusbredirhost can check for this version to check the presence of the new
usbredirhost_open_full function.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>