Add a usb_redir_filter_filter packet
Make this conditional on the same cap as device_reject, rename
the cap to usb_redir_cap_filter. Rename device_reject to filter_reject to
reflect that it is filtering related and that the the cap and packet names
are consistent with each other. Also cleanup some related comments / docs.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>