usbredirhost: Always re-attach kernel-drivers on device cleanup
Before this patch we kept track of if we actually detached the kernel
driver, and would only then re-attach the kernel driver.

But if an app is using auto-redirection there is a chance that we
claim the interface before the kernel driver got a chance to attach, this
esp. happens when the module first needs to be loaded by udev. This scenario
would leave people with USB-devices without devices attached which is very
rarely what people want! So this patch changes our behavior to instead always
re-attach the kernel drivers on device cleanup.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>