usbredirhost: Ensure we re-attach the kernel drivers after set_config failure
The set_config code path releases the interface without re-attaching the
kernel drivers. If after this the actual set_config call or the re-claim
failed the claimed flag would be 0, resulting in device cleanup not
re-attaching the kernel drivers.

So this patch keeps the claimed flag set to 1 when usbredirhost_release
only releases the interfaces without re-attaching the kernel drivers, to
ensure that they are properly re-attached on any later cleanup.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>