usbredirhost: Ack packet cancels immediately
With the upcoming qemu changes (for live migration), once a packet has been
cancelled its id may be reused for another packet immediately afterward,
the usb-guest will then expect to first get the reply to the cancelled packet,
and then to the new packet with the same id. But the actual cancellation can
finish later then the new packet.

To avoid this, this patch moves the sending back of the cancelled packet to
the usb-guest to immediately after receiving the cancel, rather then waiting
for the usb-host OS to mark the packet as completed or cancelled.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>