usbredirhost: Ensure all in flight urbs are cancelled before a reset
The guest should have cancelled any pending urbs already, but the cancellations
have not necessary completed when we receive a reset.

If we then continue with the reset pending urbs may complete with an error code
of LIBUSB_TRANSFER_NO_DEVICE, causing us to see a device disconnect.

We also need to cleanly shutdown any streams and let the guest know they
should be restarted after the reset.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>